touch exec.c, and add informetion about signal in minishell.md
