+++ Makefile
@@ -18,7 +18,9 @@ SRCS =	src/main.c	\
 		src/expansion/expand_heredoc_line.c	\
 		src/parser/parse.c	\
 		src/redirection/redirect.c	\
-		src/pipeline/pipe.c
+		src/pipeline/pipe.c	\
+		src/exec/exec.c	\
+		src/signal/signal.c
 
 OBJ_DIR = obj
 OBJS = $(SRCS:%.c=$(OBJ_DIR)/%.o)
