+++ minishell.md
@@ -78,6 +78,8 @@
 	- heredoc時のexpand
 		- delimiter(cat << EOFでいうとEOFがdelimiter)がquoteに囲まれていた場合は、heredoc時に入力されるlineに変数($から始まる変数)が入ったとしても展開されない
 		- 逆にdelimiterがquoteに囲まれていなかったら、変数はexpandされる
+	- word splitting(単語の分割)
+		- TEST="a     b         c"の場合、`echo $TEST`と実行したら`"a b c"`となり、`echo "$TEST"`と実行したら`"a     b        c"`とそのまま出力される。
 
 | 名前 | 展開・置換の意味 |
 | --- | ------------- |
@@ -90,11 +92,72 @@
 | パス名展開 |  *.txt を展開すること |
 | クォート削除 | "test.txt"の"を表示しない |
 - step12 signal
+	- 今回実装するのはCtrl + CとCtrl + \の2つ、Ctrl + Dは特に何も実装しなくても良い？
+	- Ctrl + C
+		- SIGINT（割り込みシグナル）を現在のフォアグラウンドプロセスに送信します。通常は、実行中のプロセスを中断するために使用されます。
+	- Ctrl + D
+		- これは EOF（End-Of-File）を意味します。キーボードからの入力が終了したことをシェルやプログラムに通知します。bashシェル(プロンプト)でCtrl + Dを押すと、シェルは入力が終了したと判断し、ログアウト（シェルの終了）します。ただし、プログラムが標準入力から読み込みを待っている場合(heredocやcatコマンドなど)、そのプログラムは入力の終了を検知します。
+	- Ctrl + \
+		- SIGQUIT（終了シグナル）を現在のフォアグラウンドプロセスに送信します。デフォルトの挙動としては、プロセスを終了させるだけでなく、コアダンプ（メモリの内容を含むデバッグ情報）を生成します。
+	- シグナル受信時の動作 (シグナルアクション) 
+		- シグナル受信時の動作 (シグナルアクション) には、次の 3 種類があります。この動作は後述の sigaction() などから変更できます。ただし、SIGKILL と SIGSTOP は標準動作以外にはできません。
+		1. 無視
+			- シグナルを受信しても何もしません。
+			- (設定するには、後述の sigaction() などで sa_handler に SIG_IGN を指定する)
+		1. 標準動作
+			- シグナル受信すると、シグナル毎に定義された標準動作 (後述の Term, Ign, Core, Stop, Cont) を実行します。
+			- (設定するには、後述の sigaction() などで sa_handler に SIG_DFL を指定 (デフォルト) する)
+			- 標準動作は、シグナル毎に定義されており、次の 5 種類があります。
+				1. Term プロセス終了
+				1. Ign 何もしない (sigaction() で設定できる SIG_IGN と同じ)
+				1. Core プロセス終了とコアダンプ生成 2 
+				1. Stop プロセス一時停止 (TASK_STOPPED 状態に遷移させる)
+				1. Cont 一時停止したプロセスの再開 (TASK_STOPPED 状態からの復帰)
+			- 1 から 31 番までの標準シグナルは、各々に紐づく標準動作を持ちます。(使用頻度の高いものだけ記述)
+
+| シグナル名 | シグナル番号 (x86_64) | 標準動作 | 意味 |
+| -------- | ------------------- | ------- | --- |
+| SIGINT | 2 | Term | キーボードからの割り込み (Ctrl + C) |
+| SIGQUIT | 3 | Core | キーボードからの終了 (Ctrl + Q) |
+| SIGUSR1 | 10 | Term | ユーザ定義のシグナル (その1) |
+| SIGSEGV | 11 | Core | 不正なメモリ参照 |
+| SIGUSR2 | 12 | Term | ユーザ定義のシグナル (その2) |
+| SIGTSTP | 20 | Stop | 制御端末からの停止 (Ctrl + Z) |
+
+\
+		3. シグナルハンドラ
+			- シグナル受信すると、ユーザが定義した動作を実行します。
+			- (設定するには、後述の sigaction() などで sa_handler へユーザ定義の関数を指定する)
+	- 
 - step13 environ
 - step14 builtin
 
+##doc
+- minishell
+	- [man bash](https://linuxjm.osdn.jp/html/GNU_bash/man1/bash.1.html)
+	- [minishellの作り方(usami-sanのminishell解説)](https://usatie.notion.site/minishell-29921d3ea13447ad897349acd5733d5e)
+	- [minishellの作り方で出てくるコード例のrepository](https://github.com/usatie/minishell_v2/tree/livecoding)
+	- [The Architecture of Open Source Applications (Volume 1) The Bourne-Again Shell](https://aosabook.org/en/v1/bash.html)
+	- [シェル実装の課題を完走した by jtanaka-san](https://jun-networks.hatenablog.com/entry/2021/07/02/034002)
+	- [bash再実装の課題振り返り](https://nafuka.hatenablog.com/entry/2021/03/29/194200)
+	- [Lv4minishell(1/3)](https://note.com/syamashi/n/na31efe927b41?magazine_key=m385b3208115b)
+	- [readlineとsignal by 42soul](https://intrepidgeeks.com/tutorial/minishell-readline)
+	- [minishellの概要 by 42soul](https://techdebt.tistory.com/33?category=833728)
+	- [シェルスクリプトの $ は･･･変数の接頭辞ではなく展開するときの記号(expand解説)](https://qiita.com/ko1nksm/items/1864aa2db8ce7bcf332d)
+	- [Bashの$@,$#などの$から始まる特殊な変数について](https://linuxcommand.net/bash-special-parameters/)
+	- [Linux シグナルの基本と仕組み (カーネル v5.5 時点) ](https://qiita.com/Kernel_OGSun/items/e96cef5487e25517a576#3-%E3%82%B7%E3%82%B0%E3%83%8A%E3%83%AB%E7%95%AA%E5%8F%B7%E3%81%A8%E3%82%B7%E3%82%B0%E3%83%8A%E3%83%AB%E5%90%8D)
+- git
+	- [個人開発でGithubのissue,プルリクを活用する方法](https://qiita.com/usayamadausako/items/375bdae07e381745e6eb)
+	- [git clone の際に submodule の clone を忘れたときの対処法](https://qiita.com/kentarosasaki/items/3e670567c0512b9d411e)
+	- [git reset 解説](https://qiita.com/forest1/items/f7c821565a7a7d64d60f)
+	- [DiscordのチャンネルにGitHubの通知を送るには？](https://mekurun.com/tips/discord-github/)
+	- [［GitHub］一人開発でもissueベース/セルフプルリクエストを使って開発する](https://qiita.com/braveryk7/items/5208263cd06a8878f0c2)
+- vim
+	- [Vim + Markdown](https://qiita.com/iwataka/items/5355bdf03d0afd82e7a7)
+	- [42 Header Plugin](https://github.com/42Paris/42header)
 
-## Todo
+
+## Todo, Issue
 - exec filenameでstrdupしたのをまだfreeしてない
 - ~~ubuntuで動かすとreadlineのヘッダファイルが無いとerrorが出る~~
 	- done
@@ -117,18 +180,3 @@
 	- done
 		- なぜかpushしたら上手くいくようになった。原因不明
 
-
-##doc
-- minishell
-	- [man bash](https://linuxjm.osdn.jp/html/GNU_bash/man1/bash.1.html)
-	- [minishellの作り方(usami-sanのminishell解説)](https://usatie.notion.site/minishell-29921d3ea13447ad897349acd5733d5e)
-	- [minishellの作り方で出てくるコード例のrepository](https://github.com/usatie/minishell_v2/tree/livecoding)
-	- [readlineとsignal by 42soul](https://intrepidgeeks.com/tutorial/minishell-readline)
-	- [minishellの概要 by 42soul](https://techdebt.tistory.com/33?category=833728)
-	- [シェルスクリプトの $ は･･･変数の接頭辞ではなく展開するときの記号(expand解説)](https://qiita.com/ko1nksm/items/1864aa2db8ce7bcf332d)
-- git
-	- [個人開発でGithubのissue,プルリクを活用する方法](https://qiita.com/usayamadausako/items/375bdae07e381745e6eb)
-	- [git clone の際に submodule の clone を忘れたときの対処法](https://qiita.com/kentarosasaki/items/3e670567c0512b9d411e)
-	- [git reset 解説](https://qiita.com/forest1/items/f7c821565a7a7d64d60f)
-	- [DiscordのチャンネルにGitHubの通知を送るには？](https://mekurun.com/tips/discord-github/)
-	- [［GitHub］一人開発でもissueベース/セルフプルリクエストを使って開発する](https://qiita.com/braveryk7/items/5208263cd06a8878f0c2)
